[REFACTOR] extract health checks as strategy implementations
